fix: updates svelte-legacy to runes reactivity and move custom-layout to auth-layout for clarity
