+++ src/routes/forgot-password/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-  import ForgotPasswordForm from '$lib/components/custom-layout/ForgotPasswordForm.svelte';
+  import ForgotPasswordForm from '$lib/components/auth-layout/ForgotPasswordForm.svelte';
 </script>
 
 <div class="flex h-full w-full flex-row items-center justify-center p-4">
