+++ src/routes/login/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-  import LoginForm from '$lib/components/custom-layout/LoginForm.svelte';
+  import LoginForm from '$lib/components/auth-layout/LoginForm.svelte';
 </script>
 
 <div class="flex h-full w-full flex-row items-center justify-center p-4">
