+++ src/routes/reset-password/+page.svelte
@@ -1,9 +1,11 @@
 <script>
   import { page } from '$app/stores';
-  import ResetPasswordForm from '$lib/components/custom-layout/ResetPasswordForm.svelte';
+  import ResetPasswordForm from '$lib/components/auth-layout/ResetPasswordForm.svelte';
 
-  // Get token from URL parameters
-  $: token = $page.url.searchParams.get('token') || '';
+  // Using $derived for reactive computation
+  const token = $derived($page.url.searchParams.get('token') || '');
+
+  $inspect('token', token);
 </script>
 
 <div class="flex h-full w-full flex-row items-center justify-center p-4">
