+++ src/routes/signup/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-  import SignupForm from '$lib/components/custom-layout/SignupForm.svelte';
+  import SignupForm from '$lib/components/auth-layout/SignupForm.svelte';
 </script>
 
 <div class="flex h-full w-full flex-row items-center justify-center p-4">
